refactor(rentals): split confirm return and start rental flows
